feat: add ItemEditor for shared item editing
